use int for return value of FlagEncoder.getVersion

+++ core/src/main/java/com/graphhopper/routing/util/Bike2WeightFlagEncoder.java
@@ -56,7 +56,7 @@ public class Bike2WeightFlagEncoder extends BikeFlagEncoder
     }
 
     @Override
-    public short getVersion()
+    public int getVersion()
     {
         return 1;
     }
